# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceTracker AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrackRequest {
  /** The product name to search for */
  product: string;
}

export interface PriceResult {
  /** The e-commerce site name */
  site: string;
  /** The product title found */
  title?: string;
  /** The price as a string (e.g. "$799.99") */
  price?: string;
  /** URL to the product page */
  link?: string;
  /** Error message if extraction failed */
  error?: string;
}

export interface TrackResponse {
  /** The searched product name */
  product: string;
  results: PriceResult[];
  bestDeal?: PriceResult;
  /** Job ID for status polling */
  jobId?: string;
}

export type JobStatusStatus =
  (typeof JobStatusStatus)[keyof typeof JobStatusStatus];

export const JobStatusStatus = {
  pending: "pending",
  running: "running",
  done: "done",
  error: "error",
} as const;

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  product?: string;
  /** Human-readable progress message */
  progress?: string;
  result?: TrackResponse;
  error?: string;
}

export interface ErrorResponse {
  error: string;
}