# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceTracker AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusStatus } from "./jobStatusStatus";
import type { TrackResponse } from "./trackResponse";

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  product?: string;
  /** Human-readable progress message */
  progress?: string;
  result?: TrackResponse;
  error?: string;
}