# lib/api-zod/src/generated/types/priceResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceTracker AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceResult {
  /** The e-commerce site name */
  site: string;
  /** The product title found */
  title?: string;
  /** The price as a string (e.g. "$799.99") */
  price?: string;
  /** URL to the product page */
  link?: string;
  /** Error message if extraction failed */
  error?: string;
}