# lib/api-zod/src/generated/types/trackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceTracker AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrackRequest {
  /** The product name to search for */
  product: string;
}