# lib/api-zod/src/generated/types/trackResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PriceTracker AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PriceResult } from "./priceResult";

export interface TrackResponse {
  /** The searched product name */
  product: string;
  results: PriceResult[];
  bestDeal?: PriceResult;
  /** Job ID for status polling */
  jobId?: string;
}